+++ drivers/i2c/i2c_rcar.c
@@ -12,7 +12,7 @@
 #include <soc.h>
 #include <drivers/i2c.h>
 #include <drivers/clock_control.h>
-#include <drivers/clock_control/rcar_clock_control.h>
+#include <drivers/clock_control/renesas-cpg-mssr.h>
 
 #include <logging/log.h>
 LOG_MODULE_REGISTER(i2c_rcar);
