+++ drivers/serial/uart_rcar.c
@@ -11,7 +11,7 @@
 #include <devicetree.h>
 #include <drivers/uart.h>
 #include <drivers/clock_control.h>
-#include <drivers/clock_control/rcar_clock_control.h>
+#include <drivers/clock_control/renesas-cpg-mssr.h>
 #include <drivers/pinctrl.h>
 #include <spinlock.h>
